Finished Tasks
I was able to fix the Min/Max Camera Pitch so that the player would not be able to see below the floor. Slowed down the turning as well, and I was able to figure out the issue with the Player model not moving forward as the player tried to move foward. It was an issue with the Slime model's forward.